hotfix : Block's full id is always Long

+++ app/src/main/kotlin/be/zvz/kookie/item/ItemBlock.kt
@@ -21,7 +21,7 @@ import be.zvz.kookie.block.Block
 import be.zvz.kookie.block.BlockFactory
 
 class ItemBlock(identifier: ItemIdentifier, val block: Block) : Item(identifier, block.name) {
-    private val blockFullId: Int = block.getFullId()
+    private val blockFullId: Long = block.getFullId()
 
     override fun getBlock(clickedFace: Int?): Block = BlockFactory.fromFullBlock(blockFullId)
